Changed INumberEncoder args from long to int

+++ src/ShortUrl.Application/Interfaces/INumberEncoder.cs
@@ -1,6 +1,6 @@
 ﻿namespace ShortUrl.Application.Interfaces;
 
 public interface INumberEncoder {
-    string Encode(long number);
-    long Decode(string value);
+    string Encode(int number);
+    int Decode(string value);
 }
